refactor(settings): extract index-view layout edition into a hook
Review feedback: the persist/reorder/visibility orchestration moves from
the pages component into useIndexViewFieldsLayout under modules, the
preceding-field derivation becomes a named helper instead of nested
ternaries, the visibility toggle styling interpolates the styled
component instead of class strings, and sortFieldMetadataItemsByViewLayout
narrows to the fields it reads and gains a spec.
